Remove directory based structure
